Refactored structure
Separate directories for routes, components, etc
