+++ src/app/Components/DashboardComponent/DashboardComponent.ts
@@ -8,5 +8,4 @@ import { Component } from '@angular/core';
 
 export class DashboardComponent 
 {
-
 }
