+++ src/app/Components/LoginComponent/LoginComponent.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
-import { User } from '../Models/User';
+import { User } from '../../Models/User';
 
 @Component({
   selector: 'login',
