+++ src/app/app.module.ts
@@ -5,17 +5,13 @@ import { HttpModule } from '@angular/http';
 import { RouterModule, Routes } from '@angular/router';
 
 import { AppComponent } from './app.component';
-import { LoginComponent } from './LoginComponent/LoginComponent';
-import { RegisterComponent } from './RegisterComponent/RegisterComponent';
-import { DashboardComponent } from './DashboardComponent/DashboardComponent';
-import { LandingComponent } from './LandingComponent/LandingComponent';
+import { LoginComponent } from './Components/LoginComponent/LoginComponent';
+import { RegisterComponent } from './Components/RegisterComponent/RegisterComponent';
+import { DashboardComponent } from './Components/DashboardComponent/DashboardComponent';
+import { LandingComponent } from './Components/LandingComponent/LandingComponent';
 
-const routes: Routes = [
-  { path: 'landing', component: LandingComponent },
-  { path: 'login', component: LoginComponent },
-  { path: 'register', component: RegisterComponent },
-  { path: 'dashboard', component: DashboardComponent }
-];
+import { AppRoutes } from './Routes/AppRoutes';
+
 
 @NgModule({
   declarations: [
@@ -29,7 +25,7 @@ const routes: Routes = [
     BrowserModule,
     FormsModule,
     HttpModule,
-    RouterModule.forRoot(routes)
+    RouterModule.forRoot(AppRoutes)
   ],
   providers: [],
   bootstrap: [AppComponent]
